requirements.txt
requests
beautifulsoup4
playwright
duckdb
psycopg2-binary
python-dotenv
rapidfuzz
tldextract